feat: optimize the types
