+++ packages/vbi-agent/src/builder-agent.ts
@@ -1,7 +1,7 @@
 import { createAgentRuntime } from './runtime.js'
 import { createBuilderTools } from './tools/builder-tools.js'
 import { createToolKit } from './tools/toolkit.js'
-import type { AgentRuntimeController, VBIBuilderAgentInput } from './types.js'
+import type { AgentRuntimeController, VBIBuilderAgentInput } from './types/index.js'
 
 export const createVBIBuilderAgent = ({ model, tools = [], workspace }: VBIBuilderAgentInput): AgentRuntimeController =>
   createAgentRuntime({
