+++ packages/vbi-agent/src/types/index.ts
@@ -1,0 +1,31 @@
+export type { AgentActivity, AgentState, AgentStateListener } from './activity.js'
+export type { VBIBuilderAgentInput } from './builder.js'
+export type {
+  AgentModelConfig,
+  ModelProvider,
+  ModelStreamHandlers,
+  ModelTurnResult,
+  PendingToolCall,
+  RequiredModelConfig,
+} from './model.js'
+export type { AgentRuntimeController, RuntimeInput } from './runtime.js'
+export type { AgentScriptRuntimeInput, AgentScriptRuntimeResult } from './script.js'
+export type {
+  BuiltinSkill,
+  BuiltinSkillAsset,
+  BuiltinSkillReference,
+  FrontmatterMap,
+  MarkdownSection,
+  ParsedSkillAsset,
+  SkillAssetMeta,
+  SkillReadInput,
+  SkillSearchInput,
+} from './skills.js'
+export type { AgentTool, AgentToolKit, ToolExecutionResult, VBITool } from './tool.js'
+export type {
+  VBIAgentWorkspace,
+  VBIWorkspaceConnector,
+  VBIWorkspaceConnectorRegistration,
+  VBIWorkspaceConnectors,
+  VBIWorkspaceSlot,
+} from './workspace.js'
