+++ packages/vbi-agent/src/types/script.ts
@@ -1,0 +1,9 @@
+export interface AgentScriptRuntimeInput {
+  code: string
+  globals?: Record<string, unknown>
+}
+
+export interface AgentScriptRuntimeResult {
+  logs: string[]
+  result: unknown
+}
